backend_fastapi/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.0
httpx==0.27.2
python-jose==3.3.0
passlib[bcrypt]==1.7.4
